+++ server/src/hooks/index.ts
@@ -2,3 +2,4 @@ export { default as useToken } from "./useToken";
 export { default as useAuth } from "./useAuth";
 export { default as useMarkdown } from "./useMarkdown";
 export { default as useValidPath } from "./useValidPath";
+export { default as usePermission } from "./usePermission";